- unified connection selection with Console view
