--- src/com/intellij/idea/plugin/hybris/tools/remote/console/actions/HybrisChooseInstanceAction.kt
@@ -1,53 +1,0 @@
-/*
- * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package com.intellij.idea.plugin.hybris.tools.remote.console.actions
-
-import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
-import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
-import com.intellij.idea.plugin.hybris.settings.options.ProjectIntegrationsSettingsConfigurableProvider.SettingsConfigurable
-import com.intellij.openapi.actionSystem.ActionUpdateThread
-import com.intellij.openapi.actionSystem.AnAction
-import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.actionSystem.CommonDataKeys
-import com.intellij.openapi.options.ShowSettingsUtil
-
-class HybrisChooseInstanceAction : AnAction(
-    message("action.choose.hybris.instance.message.text"),
-    message("action.choose.hybris.instance.message.title"),
-    HybrisIcons.Y.REMOTE
-) {
-
-    override fun actionPerformed(e: AnActionEvent) {
-        val project = getEventProject(e) ?: return
-
-        ShowSettingsUtil.getInstance().showSettingsDialog(project, SettingsConfigurable::class.java)
-    }
-
-    override fun update(e: AnActionEvent) {
-        val project = e.getData(CommonDataKeys.PROJECT)
-        if (project == null) {
-            e.presentation.isEnabledAndVisible = false
-            return
-        }
-
-        e.presentation.isEnabledAndVisible = true
-    }
-
-    override fun getActionUpdateThread() = ActionUpdateThread.BGT
-}
